requirements.txt
flask>=3.1.1
python-telegram-bot[job-queue]==22.3
python-dotenv==1.0.1